Revert "Revert "Updating longitude from Int to float""
This reverts commit 6acc67b8b9c10413ec4ed12a621ccf3dd5fffa56.

+++ models/place.py
@@ -16,7 +16,7 @@ class Place(BaseModel, Base):
     max_guest = Column(Integer, nullable=False, default=0)
     price_by_night = Column(Integer, nullable=False, default=0)
     latitude = Column(Float, nullable=True)
-    longitude = Column(Integer, nullable=True)
+    longitude = Column(Float, nullable=True)
     amenity_ids = []
 
     @property
